+++ src/Motus.GH/Components/MotusStewartComponent.cs
@@ -27,6 +27,8 @@ public sealed class MotusStewartComponent : RobotSourceComponentBase
     {
     }
 
+    public override GH_Exposure Exposure => GH_Exposure.secondary;
+
     protected override void RegisterInputParams(GH_InputParamManager p)
     {
         p.AddTextParameter(